backend/requirements.txt
fastapi
uvicorn
langchain
langchain-community
langchain-openai
langchain-chroma
chromadb
python-multipart
pypdf
tiktoken
python-dotenv
beautifulsoup4
